+++ zrt/src/models/word_count_stats.rs
@@ -2,10 +2,10 @@
 
 #[derive(Debug, Default)]
 pub struct WordCountStats {
-    pub tagged_files: u64,
-    pub tagged_words: u64,
-    pub total_files: u64,
-    pub total_words: u64,
+    pub tagged_files: u32,
+    pub tagged_words: u32,
+    pub total_files: u32,
+    pub total_words: u32,
 }
 
 impl WordCountStats {
@@ -21,12 +21,10 @@ impl WordCountStats {
     }
     #[inline]
     #[must_use]
-    #[expect(clippy::as_conversions, reason = "Precision not critical")]
-    #[expect(clippy::cast_precision_loss, reason = "Precision not critical")]
     pub fn calculate_percentage(&self) -> f64 {
         if self.total_words == 0 {
             return 0.0;
         }
-        (self.tagged_words as f64 / self.total_words as f64) * 100.0
+        (f64::from(self.tagged_words) / f64::from(self.total_words)) * 100.0
     }
 }
